kedro/requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro~=0.19.9
kedro-datasets>=3.0; python_version >= "3.9"
kedro-datasets>=1.0; python_version < "3.9"
kedro-viz>=6.7.0
kedro[jupyter]
notebook
scikit-learn~=1.5.1; python_version >= "3.9"
scikit-learn<=1.4.0,>=1.0; python_version < "3.9"
setuptools; python_version >= "3.12"
pyspark
seaborn
matplotlib
iterative-stratification
kedro
pandas
ipykernel
python-vlc
keyboard
scikit-learn
